pyproject.toml
[tool.poetry]
name = "polly loan calculator"
version = "0.1.0"
description = "Interview assignment"
authors = ["Piotr Gryko <piotr.gryko@gmail.com>"]
license = "proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.2"
django-ninja = "^1.3.0"
python-decouple = "^3.8"
django-cors-headers = "^4.5.0"
django-extensions = "^3.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
ruff = "^0.7.0"
black = "^24.10.0"
pylama = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
